Split views.py into multiple files
