+++ comps/views/__init__.py
@@ -1,0 +1,12 @@
+from .all_comps import *
+from .combine_heats import *
+from .comp_detail import *
+from .comp_heats import *
+from .create_comp import *
+from .fix_couple_type import *
+from .fix_null_entries import *
+from .heat import *
+from .load_dancers import *
+from .load_heats import *
+from .resolve_dancers import *
+from .resolve_mismatches import *
